backend/requirements.txt
Django>=5.2,<6
djangorestframework>=3.16
django-cors-headers>=4.9
gunicorn>=21.2
Pillow>=11.0
pytesseract>=0.3.13
